// frontend/src/models/Repository/ExoplanetDiscMethod.ts
export enum DiscoveringMethod {
  Unfiltered = "", // Unfiltered
  TRAN = "Transit", // Transit
  AST = "Astrometry",   // Astrometry
  PUL = "Pulsar Timing",   // Pulsar Timing
  OBM = "Orbital Brightness Modulation",   // Orbital Brightness Modulation
  MICRO = "Microlensing", // Microlensing
  ETV = "Eclipse Timing Variations",   // Eclipse Timing Variations
  TTV = "Transit Timing Variations",   // Transit Timing Variations
  RV = "Radial Velocity",     // Radial Velocity
  IMA = "Imaging",   // Imaging
  DKIN = "Disk Kinematics", // Disk Kinematics
  PTV = "Pulsation Timing Variations"    // Pulsation Timing Variations
}